+++ app/[locale]/template/[templateId]/page.tsx
@@ -1,10 +1,10 @@
 // template/[templateId]/section/page.tsx
 'use client';
 
-import React, { useEffect, useRef, useState } from 'react';
-import { useTranslations } from 'next-intl';
-import { useParams } from 'next/navigation';
-import { ApolloError } from "@apollo/client";
+import React, {useEffect, useRef, useState} from 'react';
+import {useTranslations} from 'next-intl';
+import {useParams} from 'next/navigation';
+import {ApolloError} from "@apollo/client";
 import {
   Breadcrumb,
   Breadcrumbs,
@@ -41,9 +41,9 @@ import AddQuestionButton from "@/components/AddQuestionButton";
 import AddSectionButton from "@/components/AddSectionButton";
 import ErrorMessages from '@/components/ErrorMessages';
 
-import { useFormatDate } from '@/hooks/useFormatDate';
+import {useFormatDate} from '@/hooks/useFormatDate';
 import logECS from '@/utils/clientLogger';
-import { useToast } from '@/context/ToastContext';
+import {useToast} from '@/context/ToastContext';
 import styles from './templateEditPage.module.scss';
 
 interface QuestionsInterface {
@@ -224,7 +224,6 @@ const TemplateEditPage: React.FC = () => {
         },
       });
     }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [data]);
 
   // If page-level errors that we need to scroll to